requirements.txt
Flask>=2.0
pymongo>=4.0
certifi
dnspython
cryptography
